Add hierarchical configuration (#4)
This commit adds clap to parse configuration options from the command
line and environment variables. Because of this, the MSRV is bumped up
to 1.70.

This is a completely backwards-compatible change, since all
configuration options can be specified in `config.toml` without any
overrides.

This also sets us up for releasing static binaries once `tower` adds
support for virtual filesystems or embedded byte slices in `ServeDir`.