remove x86_64 android target
android *really* seems to hate running in 64-bit mode,
and the NDK does not include libunwind,
which rust needs despite panic=abort, which is a bummer.

but we still build for like 20 targets, so that's fine.

This is good enough for me to start working on an installer
of some sort.